Enhance training notebook and fix copy mechanism for AMP compatibility
- Added execution counts and outputs to code cells in the training notebook for improved tracking in Colab.
- Implemented a fix in the copy mechanism to ensure compatibility with AMP by adjusting data types for scatter_add_ operations.
- Updated the notebook structure to include necessary comments and organization for better clarity.

+++ ptx_decompiler/model/copy_mechanism.py
@@ -48,6 +48,8 @@ class CopyGenerator(nn.Module):
         enc_ids = encoder_token_ids.unsqueeze(1).expand(B, T, -1)
         valid = (enc_ids >= 0) & (enc_ids < self.vocab_size)
         enc_ids_clamped = enc_ids.clamp(0, self.vocab_size - 1).long()
-        contribution = (contribution * valid.float()).to(logits_vocab.dtype)
+        contribution = contribution * valid.to(contribution.dtype)
+        # Ensure same dtype as logits_vocab for scatter_add_ (required under AMP when logits are float16)
+        contribution = contribution.to(logits_vocab.dtype)
         logits_vocab.scatter_add_(2, enc_ids_clamped, contribution)
         return logits_vocab, p_gen
